use zod in scripts
