--- libs/scripts/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@modtree/scripts",
-  "version": "0.0.1",
-  "type": "commonjs"
-}
