--- libs/scripts/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "strict": true,
-    "outDir": "../../dist",
-    "module": "commonjs",
-    "types": ["node"]
-  },
-  "files": [],
-  "include": ["**/*.ts"],
-  "exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts"]
-}
