--- libs/scripts/webpack.js
@@ -1,40 +1,0 @@
-/**
- * executing this file bundles the the server project and all its dependencies
- * into one executable pure nodejs file.
- */
-const webpack = require('webpack')
-const { resolve } = require('path')
-const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin')
-
-/**
- * configuration paths
- */
-const rootDir = resolve(__dirname, '../..')
-const outDir = resolve(rootDir, 'dist/libs/scripts')
-const tsconfig = resolve(__dirname, 'tsconfig.json')
-
-/**
- * webpack config
- */
-const compiler = webpack({
-  target: 'node',
-  mode: 'development',
-  entry: {
-    'docker-build': resolve(__dirname, 'src/docker-build'),
-  },
-  output: {
-    filename: '[name].js',
-    libraryTarget: 'commonjs-module',
-    path: outDir,
-  },
-  module: {
-    rules: [{ test: /\.ts$/, use: 'ts-loader', exclude: /node_modules/ }],
-  },
-  resolve: {
-    roots: [rootDir],
-    extensions: ['.tsx', '.ts', '.js'],
-    plugins: [new TsconfigPathsPlugin({ configFile: tsconfig })],
-  },
-})
-
-module.exports = compiler
