+++ package.json
@@ -11,6 +11,7 @@
     "build:docs": "bash scripts/docs/build.sh",
     "dev": "node scripts/serve.js",
     "build:server": "node apps/server/webpack.js --build",
+    "build:scripts": "node libs/scripts/dev.js --build",
     "start:server": "node dist/apps/server/bundle.js",
     "all:server": "yarn build:server && yarn start:server",
     "─────────────────────────────────────────────────── docker ": [],
