+++ scripts/tsconfig.json
@@ -3,11 +3,8 @@
   "compilerOptions": {
     "module": "commonjs",
     "noEmit": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "noImplicitOverride": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true
+    "declaration": true,
+    "strict": true
   },
   "files": [],
   "include": ["**/*.ts"]
